requirements.txt
python-telegram-bot
peewee
requests